Allowed for more matching files in the file ids dictionary.

+++ FinModelUtility/Fin/Fin/src/io/FileIdsDictionary.cs
@@ -80,7 +80,7 @@ public partial class FileIdsDictionary : IFileIdsDictionary {
   private partial class FileIdsPair : IBinaryConvertible {
     public uint Id { get; set; }
 
-    [SequenceLengthSource(SchemaIntegerType.BYTE)]
+    [SequenceLengthSource(SchemaIntegerType.UINT16)]
     public SizedString[] FilePaths { get; set; }
   }
